# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: blog
    description: Blog post operations
  - name: newsletter
    description: Newsletter subscriber operations
  - name: siteContent
    description: Site content management
  - name: admin
    description: Admin authentication
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /blog-posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List published blog posts
      parameters:
        - in: query
          name: all
          schema:
            type: string
          description: If "true", return all posts including drafts (admin only)
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a new blog post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostBody"
      responses:
        "201":
          description: Created blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "401":
          description: Unauthorized

  /blog-posts/{id}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a single blog post
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found
    patch:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostBody"
      responses:
        "200":
          description: Updated blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterSignupBody"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscriber"
        "409":
          description: Already subscribed

  /newsletter/subscribers:
    get:
      operationId: listNewsletterSubscribers
      tags: [newsletter]
      summary: List newsletter subscribers (admin)
      responses:
        "200":
          description: List of subscribers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsletterSubscriber"
        "401":
          description: Unauthorized

  /site-content:
    get:
      operationId: listSiteContent
      tags: [siteContent]
      summary: List all site content entries
      responses:
        "200":
          description: Site content
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SiteContent"

  /site-content/{key}:
    patch:
      operationId: updateSiteContent
      tags: [siteContent]
      summary: Update a site content entry (admin)
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteContentBody"
      responses:
        "200":
          description: Updated site content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteContent"
        "401":
          description: Unauthorized

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid password

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"

  /admin/me:
    get:
      operationId: adminMe
      tags: [admin]
      summary: Check admin authentication status
      responses:
        "200":
          description: Auth status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        tag:
          type: string
        imageUrl:
          type: ["string", "null"]
        readTime:
          type: string
        published:
          type: boolean
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - excerpt
        - content
        - category
        - tag
        - imageUrl
        - readTime
        - published
        - publishedAt
        - createdAt
        - updatedAt

    BlogPostBody:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        tag:
          type: string
        imageUrl:
          type: ["string", "null"]
        readTime:
          type: string
        published:
          type: boolean
      required:
        - title
        - excerpt
        - content
        - category
        - tag
        - readTime
        - published

    NewsletterSubscriber:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - createdAt

    NewsletterSignupBody:
      type: object
      properties:
        email:
          type: string
        name:
          type: ["string", "null"]
      required:
        - email

    SiteContent:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        value:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - key
        - value
        - updatedAt

    SiteContentBody:
      type: object
      properties:
        value:
          type: string
      required:
        - value

    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password

    AdminSession:
      type: object
      properties:
        authenticated:
          type: boolean
      required:
        - authenticated